Hvfor jan det ikke bare være simpelt en gang i mellem?
"Unfortunately, this theory has stuck around for the sole reason that it’s simple. It’s easy. People like being able to pinpoint their behaviours, particularly when it is supported by people we trust. In particular, it garnered a new wave of support after being featured in Carl Sagan’s Pulitzer Prize winner, The Dragons of Eden, despite him stressing that it wasn’t necessarily proven and it was oversimplistic. So perhaps it’s time to let go of the lizard brain and focus on what’s actually true. Don’t you think?"

https://medium.com/mind-cafe/the-science-behind-your-lizard-brain-and-why-you-cant-control-it-ca35166b68c7